refactor: factored out Compressed::decompress()
since all Compressed are also Read, we can actually provide
a blanket implementation for Compressed::decompress that is shared
for all compressed file type: Gz, Bz2 and Xz.
