+++ src/gzip.rs
@@ -1,10 +1,7 @@
-use std::fs::{create_dir_all, File};
-use std::io::{copy, Read};
-use std::path::Path;
-
 use flate2::read::GzDecoder;
+use std::{fs::File, io::Read, path::Path};
 
-use crate::{Compressed, Result};
+use crate::Compressed;
 
 pub struct Gzip<R: Read> {
     archive: GzDecoder<R>,
@@ -26,20 +23,7 @@ impl<R: Read> Gzip<R> {
     }
 }
 
-impl<R: Read> Compressed for Gzip<R> {
-    fn decompress(&mut self, target: &Path) -> Result<()> {
-        if let Some(p) = target.parent() {
-            if !p.exists() {
-                create_dir_all(p)?;
-            }
-        }
-
-        let mut output = File::create(target)?;
-        copy(&mut self.archive, &mut output)?;
-
-        Ok(())
-    }
-}
+impl<R: Read> Compressed for Gzip<R> {}
 
 impl<R: Read> Read for Gzip<R> {
     fn read(&mut self, into: &mut [u8]) -> std::io::Result<usize> {
